rework code to use JSON for settings via a script tag
rework code to use JavaScript Module Patterns for twc, twc.sppc, and twc.sppc.callbacks. makes the initialization code a bit more dense but removes the need for the user to dive into this code to set callbacks. they just need to add a callback per that codes instructions

bumped version to 0.6.0